+++ packages/client/src/components/Button/index.tsx
@@ -7,7 +7,6 @@ export type ButtonProps = {
   text: string
   view?: 'primary' | 'secondary' | 'ghost'
   className?: string
-  onClick?: () => void
 } & ButtonHTMLAttributes<HTMLButtonElement>
 
 /**
